+++ include/dataset.h
@@ -12,7 +12,7 @@ typedef struct {
   point_s array[NUM_LAYERS][MAX_NUM_POINTS];
   int n_points[NUM_LAYERS];
   size_t total_points;
-  double boundaryPoint_offset;
+  float boundaryPoint_offset;
 } dataset_s;
 
 dataset_s dataset_init(environment_s &env);
@@ -20,7 +20,7 @@ dataset_s dataset_init(environment_s &env);
 void dataset_import_data(dataset_s &dataset, point_s *data_array,
                          size_t data_array_size);
 
-void dataset_add_boundary_point(dataset_s &dataset, double offset = 0.0001);
+void dataset_add_boundary_point(dataset_s &dataset, float offset = 0.0001);
 
 // DEBUG FUNCTIONS
 std::ostream &operator<<(std::ostream &os, const dataset_s &dataset);
