+++ include/point.h
@@ -5,12 +5,12 @@
 
 typedef struct {
   int layer_num;
-  double radius;
-  double phi;
-  double z;
+  float radius;
+  float phi;
+  float z;
 } point_s;
 
-point_s point_init(int layer_num, double radius, double phi, double z);
+point_s point_init(int layer_num, float radius, float phi, float z);
 
 // DEBUG FUNCTION
 std::ostream &operator<<(std::ostream &os, const point_s &p);
